Refactor LocationSection and StorageLocationColumn components to utilize new useBucketLocationDisplay hook for improved location handling. Remove redundant code and enhance readability.
